Fix rendering of color_plain background in kmahjongg, kshisen, and libkmahjongg.
In the process, I had to remove the manual double-buffering from
KShisen's Board class, in order to make the painting work in cases
without a background brush. This is not so bad, however, because Qt has
internal double-buffering since quite some time.

Also, add a "Description" on libkmahjongg/backgrounds/color_plain.desktop

BUG: 223450

svn path=/trunk/KDE/kdegames/libkmahjongg/; revision=1167011