RKMahjonggConfigDialog: remove attempt to add another button box
KConfigDialog already adds a button box itself.
Our's stayed unused & hidden, because it was not added as child widget
and the layout with it we tried to set was rejected, as there was already
a layout installed by KConfigDialog & base classes itself. cmp. the
runtime warning
QWidget::setLayout: Attempting to set QLayout "" on
KMahjonggConfigDialog "settings", which already has a layout