Tile sets: hide design helper lines around tile "faces"
The SVG files are using a <rect> with a fine outline to mark the area
of each "face". That outline is just to assist in the design process,
not intended for ending up in the game visuals.

Thus a stroke width of 0.1 has been used originally, as that would be
still visible in the SVG editor, even more when zoomed in, but
practically invisible in the small tile graphics in the game.

With HiDPI display that no longer works, at least in the theme selector
where the one tile in the preview is shown even larger, that outline now
appears and damages the look.

The "Imperial Jade" graphics already have the outline of the area rect
set to "stroke:none" instead of "stroke:black".

So this patch follows that approach for all other themes, for some
simple approach for now.

(all besides "Ancient Egyptians" deploy <use> to just have the area rect
element defined once, so there only the original rect has to be adjusted)

(cherry picked from commit fd478fd3003aab81a90f1075d1e1305d81c4d5d2)